Related work items: #19047

+++ java/commoncomponent/daf.postgre.connection/src/main/java/net/atos/daf/postgre/bo/LiveFleetPojo.java
@@ -44,7 +44,7 @@ public class LiveFleetPojo implements Serializable{
 	private Double gpsSpeed;
 	private Long gpsDatetime; 
 	private Double wheelbasedSpeed; 
-	private Double tachgraphSpeed ;
+	private Integer tachgraphSpeed ;
 	private String driver1Id;
 	private String vehicleMsgTriggerAdditionalInfo; 
 	private Integer driverAuthEquipmentTypeId;
